Fix: named spawnP properly
